+++ examples/react-waas/src/config.ts
@@ -1,8 +1,6 @@
 import { SequenceCheckoutConfig } from '@0xsequence/checkout'
 import { ConnectConfig, createConfig } from '@0xsequence/connect'
 import { ChainId } from '@0xsequence/connect'
-import { Environment } from '@imtbl/config'
-import { passport } from '@imtbl/sdk'
 import { zeroAddress } from 'viem'
 
 const searchParams = new URLSearchParams(location.search)
@@ -71,19 +69,6 @@ export const connectConfig: ConnectConfig = {
       }
     : undefined
 }
-
-export const passportInstance = new passport.Passport({
-  baseConfig: {
-    environment: Environment.SANDBOX,
-    publishableKey: 'pk_imapik-test-VEMeW7wUX7hE7LHg3FxY'
-  },
-  forceScwDeployBeforeMessageSignature: true,
-  clientId: 'ap8Gv3188GLFROiBFBNFz77DojRpqxnS',
-  redirectUri: `${window.location.origin}/auth-callback`,
-  logoutRedirectUri: `${window.location.origin}`,
-  audience: 'platform_api',
-  scope: 'openid offline_access email transact'
-})
 
 export const config = createConfig('waas', {
   ...connectConfig,