+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @0xsequence/hooks
 
+## 6.0.4
+
+### Patch Changes
+
+- Fix for node urls
+
+## 6.0.3
+
+### Patch Changes
+
+- Remove unnecessary waas components
+
 ## 6.0.2
 
 ### Patch Changes